+++ src/app/manifest/selectors.ts
@@ -1,9 +1,13 @@
 import { destinyVersionSelector } from 'app/accounts/selectors';
 import { RootState } from 'app/store/types';
+import { emptyArray } from 'app/utils/empty';
 import { useSelector } from 'react-redux';
 
 export const destiny2CoreSettingsSelector = (state: RootState) =>
   state.manifest.destiny2CoreSettings;
+
+export const rankProgressionHashesSelector = (state: RootState) =>
+  state.manifest.destiny2CoreSettings?.currentRankProgressionHashes ?? emptyArray<number>();
 
 export const d1ManifestSelector = (state: RootState) => state.manifest.d1Manifest;
 export const d2ManifestSelector = (state: RootState) => state.manifest.d2Manifest;